+++ src/ui/pages/Portfolio/PortfolioSubmenu/SubmenuActionButtons/WithdrawDepositFlow/WithdrawMarginDialog/WithdrawMarginDialogContent/WithdrawMarginDetails/WithdrawMarginDetails.styled.tsx
