+++ src/ui/pages/Portfolio/PortfolioSubmenu/SubmenuActionButtons/WithdrawDepositFlow/WithdrawMarginDialog/WithdrawMarginSuccessDialogContent/WithdrawMarginSuccessDialogContent.styled.tsx
