+++ src/ui/pages/Portfolio/PortfolioSubmenu/SubmenuActionButtons/WithdrawDepositFlow/WithdrawMarginDialog/WithdrawMarginSuccessDialogContent/WithdrawSuccessMarginDetails/WithdrawSuccessMarginDetails.styled.tsx
